# More flexibility to optimize for individual usages