--- a/README.md
+++ b/README.md
@@ -19,8 +19,40 @@ This gem lets your API users pass in and receive camelCased or dash-cased keys, 
 
 Include a `X-Key-Inflection` header with values of `camel`, `dash`, or `snake` in your JSON API requests.
 
+
 For more examples, see [our blog post](https://www.viget.com/articles/introducing-olivebranch).
 
+## Optimizations and configuration
+
+`OliveBranch` uses `multi_json`, which will choose the fastest available JSON parsing library and use that. Combined with `Oj` can speed things up and save ~20% rails response time.
+
+The middleware can be initialized with custom camelize/dasherize implementations, so if you know you have a fixed size set of keys, you can save a considerable amount of time by providing a custom camelize that caches like so:
+
+```ruby
+class FastCamel
+  def self.camel_cache
+    @camel_cache ||= {}
+  end
+
+  def self.camelize(string)
+    camel_cache[string] ||= string.underscore.camelize(:lower)
+  end
+end
+
+
+...
+
+    config.middleware.use OliveBranch::Middleware, camelize: FastCamel.method(:camelize)
+
+```
+
+A benchmark of this compared to the standard implementation shows a saving of ~75% rails response times for a complex response payload, or a ~400% improvement, but there is a risk of memory usage ballooning if you have dynamic keys. You can make this method as complex as required, but keep in mind that it will end up being called a _lot_ in a busy app, so it's worth thinking about how to do what you need in the fastest manner possible.
+
+It is also possible to include a custom content type check in the same manner
+
+```ruby
+    config.middleware.use OliveBranch::Middleware, content_type_check: -> (content_type) { content_type == "my/content-type" }
+```
 
 * * *