--- a/Rakefile
+++ b/Rakefile
@@ -6,3 +6,7 @@ begin
   task default: :spec
 rescue LoadError
 end
+
+task :benchmark do
+  sh "BENCHMARK_REPETITIONS=1000 rspec spec/benchmark_spec.rb"
+end