--- a/olive_branch.gemspec
+++ b/olive_branch.gemspec
@@ -14,6 +14,7 @@ Gem::Specification.new do |s|
   s.license     = "MIT"
 
   s.add_dependency "rails", ">= 4.0"
+  s.add_dependency "multi_json"
 
   s.add_development_dependency "rspec", "~> 3.2"
 end