… Gemfile
… spec/test_app/example_responses/complex.json