Someone mentioned this project to me and I had a look over it. The app I work on was already doing similar things within rails, but I do really like the idea of pushing it all out to middleware and am hoping to move over to using this instead.

However we've come across and addressed some big performance gotchas in our implementation that cut out time spent in ruby in half as measured by newrelic, which I noticed are present and not possible to work around with the current implementation so I though I'd PR some tweaks

TLDR;

1000 repetitions with a complex payload against the rails test_app in the spec directory

before:
```
                 user     system      total        real
time taken 122.950000   1.600000 124.550000 (124.771544)
```

after: (with Oj plus a caching implementation of camelize)
```
                 user     system      total        real
time taken  34.800000   0.170000  34.970000 ( 34.981416)
```

The longer version:
* added `mutli_json`, so if a faster json implementation is available it is used, without adding a hard dependency on any particular json library (125s vs 96s in benchmark using Oj)
* added option to provide a custom implementation of camelize/dasherize, which can be used for things like caching, or avoiding modifying particular keys (for example uuids) (99s vs 35s in benchmark, using a caching camelize)
* added option to provide a custom content type check, since not every app that speaks json using content type `application/json`, for example [any json-api app](http://jsonapi.org/format/#content-negotiation-clients)

My apologies for the shonky benchmark "specs", I wasn't really sure of a better way to include them