@efatsi 2017-09-14
Wow, this looks really awesome @lparry. Sorry for the delayed review. Cool to see such clean functional ruby code though, thanks for bringing your work over into olive branch.

We just recently cut off the 2.0.0 release, thoughts on a 2.1.0 release?