@lparry 2017-09-19
sorry for the slack response, following semver I believe cutting a 2.1.0 would be right. There shouldn't be breaking changes but there are new options introduced