@efatsi 2017-09-19
This has been merged in and pushed out on a 2.1.0 release. I cleaned up gem dependencies (moved everything out of Gemfile and added necessary gem references to olive_branch.gemspec) but aside from that made no changes before cutting off 2.1.0. Thanks again for the work here.